FROMLIST: sched: rt scheduler sets capacity requirement
RT tasks don't provide any running constraints like deadline ones
except their running priority. The only current usable input to
estimate the capacity needed by RT tasks is the rt_avg metric. We use
it to estimate the CPU capacity needed for the RT scheduler class.

In order to monitor the evolution for RT task load, we must
peridiocally check it during the tick.

Then, we use the estimated capacity of the last activity to estimate
the next one which can not be that accurate but is a good starting
point without any impact on the wake up path of RT tasks.

BUG=chrome-os-partner:44828
TEST=Boot kernel on Oak.

Change-Id: I7acd9741adbe94a8a9763f77fb1d62de5d5eff8a
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Steve Muckle <smuckle@linaro.org>
(am from https://patchwork.kernel.org/patch/7804941/)
Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>

Signed-off-by: Eliminater74 <eliminater74@gmail.com>